I'm working on editing the tokenizer to mirror a language specification. However we want to modify the language spec is to-be-decided, I've just started adapting one from the Go programming language spec. With it mirrored, edits to the tokenizer based on edits to the grammar should be easy since it's just updating the regex in tokenizer.py that corresponds to the grammar rule.
I'd also like to simplify the lexer pass so that it identifies the token type
as it sees the token, rather than having that be a later step carried out in
typing.py. But that's TODO on my part.

We can meet up and do an edit-by-edit merge of this branch once I'm done with
the edits.